+++ src/engine/entities/traits/entity_traits.h
@@ -7,7 +7,7 @@
 /** Global entity type system data */
 namespace EntityTypeSystem
 {
-	static inline size_t TypeIDCounter = 0;
+	static inline u32 TypeIDCounter = 0;
 }
 
 /** Used to auto register a new entity type */
@@ -16,7 +16,7 @@ struct T_EntityTypeBase
 {
 	static inline Array<TraitID, EntityTraitsManager::max_traits> traits{};
 	// max instances per world chunk
-	static inline size_t instance_budget = 10;
+	static inline u32 instance_budget = 10;
 private:
 	static inline TypeID TYPE_ID;
